requirements.txt
pandas==2.3.3
streamlit==1.55.0
google-generativeai==0.1.0
geopy==2.4.0
pydeck==0.8.0

